Split API docs into pages & add sidebar
