+++ assets/scss/_variables.scss
@@ -24,6 +24,7 @@ $breakpoint: 960px;
 $mobile: 768px;
 $shadow: 0 0 $whitespace 0 rgba($black, .1);
 $shadow-bottom: 0 $whitespace $whitespace (-$whitespace) rgba($black, .1);
+$shadow-sidebar: $whitespace 0 $whitespace (-$whitespace) rgba($black, .1);
 
 $vs-border-width: 0;
 $vs-border-radius: $radius;
