+++ assets/scss/style.scss
@@ -27,6 +27,7 @@
 @import "components/json_example";
 @import "components/animated_background";
 @import "components/transition_height";
+@import "components/api";
 
 @import "views/index";
 @import "views/libraries";
@@ -35,4 +36,3 @@
 @import "views/libraries-library-tutorials";
 @import "views/libraries-library-tutorials-tutorial";
 @import "views/about";
-@import "views/api";
