+++ components/api/errors.vue
@@ -1,66 +1,51 @@
 <template>
-    <div>
-        <h2 id="errors">
-            Error responses from the cdnjs API
-        </h2>
-        <div class="doc">
-            <p>
-                All errors generated by using the cdnjs API will return with a unified structure for easy detection
-                &amp; processing. They will return a JSON object containing the error status code as well as a message
-                explaining the error. The HTTP response code will also reflect the error.
-            </p>
-            <p>
-                404 error responses will have a 1 hour cache lifetime, all other errors (5xx etc.) have no cache
-                lifetime set.
-            </p>
+    <div class="doc">
+        <h3>JSON Response</h3>
+        <table>
+            <thead>
+                <tr>
+                    <th>Property</th>
+                    <th>Type</th>
+                    <th>Description</th>
+                </tr>
+            </thead>
+            <tbody>
+                <tr v-for="field in response">
+                    <td><code class="key">{{ field.prop }}</code></td>
+                    <td>
+                        <template v-for="type in field.type">
+                            <code class="key">{{ type }}</code>&nbsp;
+                        </template>
+                    </td>
+                    <td>
+                        <p v-for="desc in field.desc" v-html="desc"></p>
+                    </td>
+                </tr>
+            </tbody>
+        </table>
 
-            <h3>JSON Response</h3>
-            <table>
-                <thead>
-                    <tr>
-                        <th>Property</th>
-                        <th>Type</th>
-                        <th>Description</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    <tr v-for="field in response">
-                        <td><code class="key">{{ field.prop }}</code></td>
-                        <td>
-                            <template v-for="type in field.type">
-                                <code class="key">{{ type }}</code>&nbsp;
-                            </template>
-                        </td>
-                        <td>
-                            <p v-for="desc in field.desc" v-html="desc"></p>
-                        </td>
-                    </tr>
-                </tbody>
-            </table>
+        <h3>Example Requests</h3>
 
-            <h3>Example Requests</h3>
+        <h4>
+            <a href="https://api.cdnjs.com/libraries/this-library-doesnt-exist">
+                <code class="key">https://api.cdnjs.com/libraries/this-library-doesnt-exist</code>
+            </a>
+        </h4>
+        <JSONExample>{"error":true,"status":404,"message":"Library not found"}</JSONExample>
 
-            <h4>
-                <a href="https://api.cdnjs.com/libraries/this-library-doesnt-exist">
-                    <code class="key">https://api.cdnjs.com/libraries/this-library-doesnt-exist</code>
-                </a>
-            </h4>
-            <JSONExample>{"error":true,"status":404,"message":"Library not found"}</JSONExample>
+        <h4>
+            <a href="https://api.cdnjs.com/libraries/backbone.js/tutorials/this-tutorial-doesnt-exist">
+                <code class="key">https://api.cdnjs.com/libraries/backbone.js/tutorials/this-tutorial-doesnt-exist</code>
+            </a>
+        </h4>
+        <JSONExample>{"error":true,"status":404,"message":"Tutorial not found"}</JSONExample>
 
-            <h4>
-                <a href="https://api.cdnjs.com/libraries/backbone.js/tutorials/this-tutorial-doesnt-exist">
-                    <code class="key">https://api.cdnjs.com/libraries/backbone.js/tutorials/this-tutorial-doesnt-exist</code>
-                </a>
-            </h4>
-            <JSONExample>{"error":true,"status":404,"message":"Tutorial not found"}</JSONExample>
-
-            <h4>
-                <a href="https://api.cdnjs.com/this-route-doesnt-exist">
-                    <code class="key">https://api.cdnjs.com/this-route-doesnt-exist</code>
-                </a>
-            </h4>
-            <JSONExample>{"error":true,"status":404,"message":"Endpoint not found"}</JSONExample>
-        </div>
+        <h4>
+            <a href="https://api.cdnjs.com/this-route-doesnt-exist">
+                <code class="key">https://api.cdnjs.com/this-route-doesnt-exist</code>
+            </a>
+        </h4>
+        <JSONExample>{"error":true,"status":404,"message":"Endpoint not found"}</JSONExample>
     </div>
 </template>
