+++ pages/api/errors.vue
@@ -1,0 +1,62 @@
+<template>
+    <section>
+        <APISidebar></APISidebar>
+        <div class="docs-main">
+            <header>
+                <Breadcrumbs :crumbs="crumbs"></Breadcrumbs>
+                <div class="content">
+                    <h1>
+                        Error responses from the cdnjs API
+                    </h1>
+                    <p>
+                        All errors generated by using the cdnjs API will return with a unified structure for easy detection
+                        &amp; processing. They will return a JSON object containing the error status code as well as a message
+                        explaining the error. The HTTP response code will also reflect the error.
+                    </p>
+                    <p>
+                        404 error responses will have a 1 hour cache lifetime, all other errors (5xx etc.) have no cache
+                        lifetime set.
+                    </p>
+                </div>
+            </header>
+            <div class="content">
+                <APIErrors></APIErrors>
+            </div>
+        </div>
+    </section>
+</template>
+
+<script>
+import Breadcrumbs from '../../components/breadcrumbs';
+import APISidebar from '../../components/api/sidebar';
+import APIErrors from '../../components/api/errors';
+import breadcrumbs from '../../util/breadcrumbs';
+import setMeta from '../../util/set_meta';
+
+const meta = {
+    title: 'Error responses - API Documentation',
+    breadcrumb: 'Error responses',
+    desc: 'Error responses from the cdnjs API - The cdnjs API allows for easy programmatic navigation of our libraries and tutorials. Read the docs to learn about all the endpoints we offer.',
+    classes: [
+        'api-docs',
+    ],
+};
+
+export default {
+    name: 'API',
+    meta,
+    components: {
+        Breadcrumbs,
+        APISidebar,
+        APIErrors,
+    },
+    async asyncData (data) {
+        return {
+            crumbs: await breadcrumbs(data.route, data.app.router, {}),
+        };
+    },
+    head () {
+        return setMeta(meta, this);
+    },
+};
+</script>
