+++ pages/api/library.vue
@@ -1,0 +1,60 @@
+<template>
+    <section>
+        <APISidebar></APISidebar>
+        <div class="docs-main">
+            <header>
+                <Breadcrumbs :crumbs="crumbs"></Breadcrumbs>
+                <div class="content">
+                    <h1>
+                        Getting a specific library on cdnjs
+                    </h1>
+                    <p>
+                        The <code>/libraries/:library</code> endpoint allows for data on a specific library to be
+                        requested and will return a JSON object with all library data properties by default.
+                    </p>
+                    <p>
+                        The cache lifetime on this endpoint is six hours.
+                    </p>
+                </div>
+            </header>
+            <div class="content">
+                <APILibrary></APILibrary>
+            </div>
+        </div>
+    </section>
+</template>
+
+<script>
+import Breadcrumbs from '../../components/breadcrumbs';
+import APISidebar from '../../components/api/sidebar';
+import APILibrary from '../../components/api/library';
+import breadcrumbs from '../../util/breadcrumbs';
+import setMeta from '../../util/set_meta';
+
+const meta = {
+    title: 'Get library - API Documentation',
+    breadcrumb: 'Get library',
+    desc: 'Getting a specific library on cdnjs - The cdnjs API allows for easy programmatic navigation of our libraries and tutorials. Read the docs to learn about all the endpoints we offer.',
+    classes: [
+        'api-docs',
+    ],
+};
+
+export default {
+    name: 'API',
+    meta,
+    components: {
+        Breadcrumbs,
+        APISidebar,
+        APILibrary,
+    },
+    async asyncData (data) {
+        return {
+            crumbs: await breadcrumbs(data.route, data.app.router, {}),
+        };
+    },
+    head () {
+        return setMeta(meta, this);
+    },
+};
+</script>
